requirements.txt
fastapi==0.70.1
numpy==1.19.5
Pillow==8.4.0
pydantic==1.8.2
tensorflow-cpu==2.6.2
tensorflow_hub==0.12.0
uvicorn==0.15.0
